--- src/components/MyPage/TestLoading.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { BeatLoader } from 'react-spinners';
-
-type Props = {};
-
-const TestLoading = (props: Props) => {
-  return (
-    <div>
-      <BeatLoader color='#2ac1bc' />
-    </div>
-  );
-};
-
-export default TestLoading;
